requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==3.0.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audio-extract==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icecream==2.1.3
ics==0.7.2
idna==3.7
imageio-ffmpeg==0.4.8
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
motor==3.3.2
mpmath==1.3.0
multidict==6.0.5
mutagen==1.46.0
networkx==3.3
ngrok==1.0.0
numpy==1.26.4
openai==1.3.3
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.0
pytz==2022.2.1
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
redmail==0.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.2
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
asyncpg==0.29.0